fix(token)!: stop TokenMetadata reporting fields it cannot hold
`TokenMetadata` carried two fields that were structurally always 0, so the
struct described something the token could not represent:

  * `game_id: u64` — a registry id. The registry generation is retired and the
    id layout has no such field, so `to_token_metadata` hardcoded 0.
  * `metadata: u16` — the id packs 65 bits here. The struct could not hold
    them, so rather than truncate, `to_token_metadata` hardcoded 0 and
    `mint_metadata()` existed to expose the real value.

Both were pinned by the same constraint, stated in the old doc comment: the
struct was shared with the deployed legacy token, whose ABI could not change.
That constraint died with the registry generation in v2.1.0, so the fields can
now say what is true.

  * `game_id` is removed.
  * `metadata` widens to `u128` and carries the packed value, so
    `token_metadata(id).metadata == mint_metadata(id)`. `mint_metadata` stays
    as a cheap single-field read, like `settings_id` and `objective_id`.

THE SAME WIDENING, FINISHED

d4aaac8 widened metadata to 65 bits but only touched the token interface.
The minigame-facing mint path was left at `u16`, so a caller minting through
`IMinigame::mint_game` or `mint_game_batch` could set 16 of the 65 bits while
the token's own `mint` accepted all of them — the same field, two entrypoints,
two widths. `MintGameParams.metadata` and `mint_game`'s parameter are now
`u128`, matching `MintMetagameParams`, which was already widened.
`MintParams.metadata` follows for parity.

The renderer drops the always-0 "Game ID" trait and emits the real metadata
instead of a widened u16.

INTERFACE IDS

Both move, since SRC5 extended selectors expand parameter and return types:

  IMINIGAME_ID        0x3d1730c2... -> 0x1b78fcd155ca1cfe04a0f0f75dd48b398995d2f9ff0c9f0e00ec80c71d1f2bb
  IMINIGAME_TOKEN_ID  0x15951d6d... -> 0x20253de95bcdb23620c88405a5f97da040b91de832ad98a34b45c4f3331d13b

Both rederived with src5_rs, and the method validated first by reproducing the
OLD constants from the unmodified traits. IMINIGAME_TOKEN_ID keeps the
established refresh-exclusion convention — derived over the surface minus
`refresh_metadata` — so only the token_metadata selector changed within it.
Docs carrying the literal id are updated with it.

Version 2.1.0 -> 2.2.0: v2.1.0 is released, and this changes the ABI.

Full matrix: 1318 passed. The 2 failures in
`utilities::distribution::test_payout` are pre-existing — verified by stashing
this change and reproducing them on a clean tree. CI's test-compilation
warning gate is clean.

NOTE FOR CONSUMERS

Anything decoding a token id keeps working — the id layout is untouched. What
changes is the ABI: contracts implementing `IMinigame::mint_game` must widen
their `metadata` parameter, and anything registering or probing the two
interface ids above must use the new values.

+++ packages/interfaces/src/structs/minigame.cairo
@@ -53,7 +53,10 @@ pub struct MintGameParams {
     pub soulbound: bool,
     pub paymaster: bool,
     pub salt: u16,
-    pub metadata: u16,
+    /// Inert data the game interprets. `u128` to reach the token id's 65-bit
+    /// field — a `u16` here would cap this mint path at 16 of those bits
+    /// while the token's own `mint` accepted all 65.
+    pub metadata: u128,
 }
 
 #[derive(Drop, Serde, Clone)]
